complete func:fetch movie info from web

+++ preparation/datatomysql.py
@@ -37,35 +37,81 @@ class DataToSql():
             raise e
             self.logger.error(str(e))
 
-    def create_database(self,databasename):
+    def create_database(self,databasename,ifexist="fail"):
+        '''
+
+        :param databasename:
+        :param ifexist: "fail",如果存在则报错
+                        "append",如果存在，则用当前数据库
+                        "replace",如果存在，删除当前数据库，新建同名数据库
+        :return:
+        '''
         try:
             # 新建MovieRecommender数据库
-            sql_creatDatabase = ["drop database if exists MovieRecommender;",
-                                 "create database MovieRecommender;",
-                                 "use MovieRecommender;",
-                                 ]
-            for sql in sql_creatDatabase:
-                self.cur.execute(sql)
+            if ifexist == "fail":
+                sql_creatDatabase = [f"create database {databasename};"
+                                     ]
+                for sql in sql_creatDatabase:
+                    self.cur.execute(sql)
 
-            self.logger.info('成功创建MovieRecommender数据库')
+
+            elif ifexist == "append":
+                sql_creatDatabase = [f"SELECT * FROM information_schema.SCHEMATA where SCHEMA_NAME='{databasename}';",
+                                     f"create database {databasename};"
+                                     ]
+                res = self.cur.execute(sql_creatDatabase[0])
+                if res == 0:#数据库不存在
+                    self.cur.execute(sql_creatDatabase[1])
+                else:
+                    self.logger.warning(f'{databasename}已存在')
+
+
+
+
+            else:
+                sql_creatDatabase = ["drop database if exists MovieRecommender;",
+                                     f"create database {databasename};",
+                                     f"use {databasename};",
+                                     ]
+                for sql in sql_creatDatabase:
+                    self.cur.execute(sql)
+
+                self.logger.info(f'成功新建数据库{databasename}，如果已存在，则覆盖')
 
         except:
             raise Exception
-            logger.error('MovieRecommender数据库创建失败')
+
+            self.logger.error(f'{databasename}数据库创建失败')
 
 
-    #把links.csv中的数据导入到mysql中
-    try:
-        tosql.csvtosql('../data/links.csv','MovieRecommender',)
-        logger.info('links.csv导入数据库成功')
 
-    except Exception as e:
-        print(e)
-        logger.error('links.csv导入数据库失败'+ str(e))
+
+    def push_data_tosql(self,filepath,databasename,table_name):
+        #把links.csv中的数据导入到mysql中
+        try:
+            tosql.csvtosql(filepath=filepath,database=databasename,header=0,table_name=table_name)
+            self.logger.info(f'{filepath}导入数据库成功')
+
+        except Exception as e:
+            raise e
+            self.logger.error(f'{filepath}导入数据库失败')
+
+    def close_connect(self):
+        self.cur.close()
+        self.conn.close()
+        return
+
+
 
 
 
 if __name__ == '__main__':
-    datatosql()
+    datatosql = DataToSql()
+    #datatosql.create_database('MovieRecommender',ifexist='append')
+    #datatosql.push_data_tosql('../data/links.csv','MovieRecommender','links')
+    datatosql.push_data_tosql('../data/movies.csv','MovieRecommender','movies')
+    datatosql.push_data_tosql('../data/ratings.csv','MovieRecommender','ratings')
+    datatosql.push_data_tosql('../data/users.csv','MovieRecommender','users')
+    datatosql.close_connect()
